third_party/blink/web_tests/external/wpt/css/css-text-decor/animations/text-decoration-thickness-interpolation-expected.txt

This is a testharness.js-based test.
Found 116 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSS Transitions: property <text-decoration-thickness> from [16px] to [0px] at (0) should be [16px]
  assert_equals: expected "16px " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [16px] to [0px] at (0.3) should be [11.2px]
  assert_equals: expected "11.2px " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [16px] to [0px] at (0.6) should be [6.4px]
  assert_equals: expected "6.4px " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [16px] to [0px] at (0) should be [16px]
  assert_equals: expected "16px " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [16px] to [0px] at (0.3) should be [11.2px]
  assert_equals: expected "11.2px " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [16px] to [0px] at (0.6) should be [6.4px]
  assert_equals: expected "6.4px " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [16px] to [32px] at (0) should be [16px]
  assert_equals: expected "16px " but got "32px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [16px] to [32px] at (0.3) should be [20.8px]
  assert_equals: expected "20.8px " but got "32px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [16px] to [32px] at (0.6) should be [25.6px]
  assert_equals: expected "25.6px " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [16px] to [32px] at (0) should be [16px]
  assert_equals: expected "16px " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [16px] to [32px] at (0.3) should be [20.8px]
  assert_equals: expected "20.8px " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [16px] to [32px] at (0.6) should be [25.6px]
  assert_equals: expected "25.6px " but got "32px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [1em] to [0em] at (0) should be [16px]
  assert_equals: expected "16px " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [1em] to [0em] at (0.3) should be [11.2px]
  assert_equals: expected "11.2px " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [1em] to [0em] at (0.6) should be [6.4px]
  assert_equals: expected "6.4px " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [1em] to [0em] at (0) should be [16px]
  assert_equals: expected "16px " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [1em] to [0em] at (0.3) should be [11.2px]
  assert_equals: expected "11.2px " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [1em] to [0em] at (0.6) should be [6.4px]
  assert_equals: expected "6.4px " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [1em] to [2em] at (0) should be [16px]
  assert_equals: expected "16px " but got "32px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [1em] to [2em] at (0.3) should be [20.8px]
  assert_equals: expected "20.8px " but got "32px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [1em] to [2em] at (0.6) should be [25.6px]
  assert_equals: expected "25.6px " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [1em] to [2em] at (0) should be [16px]
  assert_equals: expected "16px " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [1em] to [2em] at (0.3) should be [20.8px]
  assert_equals: expected "20.8px " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [1em] to [2em] at (0.6) should be [25.6px]
  assert_equals: expected "25.6px " but got "32px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [0%] at (0) should be [100%]
  assert_equals: expected "100 % " but got "0 % "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [0%] at (0.3) should be [70%]
  assert_equals: expected "70 % " but got "0 % "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [0%] at (0.6) should be [40%]
  assert_equals: expected "40 % " but got "0 % "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [0%] at (0) should be [100%]
  assert_equals: expected "100 % " but got "0 % "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [0%] at (0.3) should be [70%]
  assert_equals: expected "70 % " but got "0 % "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [0%] at (0.6) should be [40%]
  assert_equals: expected "40 % " but got "0 % "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [200%] at (0) should be [100%]
  assert_equals: expected "100 % " but got "200 % "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [200%] at (0.3) should be [130%]
  assert_equals: expected "130 % " but got "200 % "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [200%] at (0.6) should be [160%]
  assert_equals: expected "160 % " but got "200 % "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [200%] at (0) should be [100%]
  assert_equals: expected "100 % " but got "200 % "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [200%] at (0.3) should be [130%]
  assert_equals: expected "130 % " but got "200 % "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [200%] at (0.6) should be [160%]
  assert_equals: expected "160 % " but got "200 % "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [16px] to [0em] at (0) should be [16px]
  assert_equals: expected "16px " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [16px] to [0em] at (0.3) should be [11.2px]
  assert_equals: expected "11.2px " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [16px] to [0em] at (0.6) should be [6.4px]
  assert_equals: expected "6.4px " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [16px] to [0em] at (0) should be [16px]
  assert_equals: expected "16px " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [16px] to [0em] at (0.3) should be [11.2px]
  assert_equals: expected "11.2px " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [16px] to [0em] at (0.6) should be [6.4px]
  assert_equals: expected "6.4px " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [16px] to [2em] at (0) should be [16px]
  assert_equals: expected "16px " but got "32px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [16px] to [2em] at (0.3) should be [20.8px]
  assert_equals: expected "20.8px " but got "32px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [16px] to [2em] at (0.6) should be [25.6px]
  assert_equals: expected "25.6px " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [16px] to [2em] at (0) should be [16px]
  assert_equals: expected "16px " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [16px] to [2em] at (0.3) should be [20.8px]
  assert_equals: expected "20.8px " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [16px] to [2em] at (0.6) should be [25.6px]
  assert_equals: expected "25.6px " but got "32px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [16px] to [0%] at (0) should be [calc(0% + 16px)]
  assert_equals: expected "calc ( 0 % + 16px ) " but got "0 % "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [16px] to [0%] at (0.3) should be [calc(0% + 11.2px)]
  assert_equals: expected "calc ( 0 % + 11.2px ) " but got "0 % "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [16px] to [0%] at (0.6) should be [calc(0% + 6.4px)]
  assert_equals: expected "calc ( 0 % + 6.4px ) " but got "0 % "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [16px] to [0%] at (0) should be [calc(0% + 16px)]
  assert_equals: expected "calc ( 0 % + 16px ) " but got "0 % "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [16px] to [0%] at (0.3) should be [calc(0% + 11.2px)]
  assert_equals: expected "calc ( 0 % + 11.2px ) " but got "0 % "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [16px] to [0%] at (0.6) should be [calc(0% + 6.4px)]
  assert_equals: expected "calc ( 0 % + 6.4px ) " but got "0 % "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [16px] to [200%] at (0) should be [calc(0% + 16px)]
  assert_equals: expected "calc ( 0 % + 16px ) " but got "200 % "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [16px] to [200%] at (0.3) should be [calc(60% + 11.2px)]
  assert_equals: expected "calc ( 60 % + 11.2px ) " but got "200 % "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [16px] to [200%] at (0.6) should be [calc(120% + 6.4px)]
  assert_equals: expected "calc ( 120 % + 6.4px ) " but got "200 % "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [16px] to [200%] at (0) should be [calc(0% + 16px)]
  assert_equals: expected "calc ( 0 % + 16px ) " but got "200 % "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [16px] to [200%] at (0.3) should be [calc(60% + 11.2px)]
  assert_equals: expected "calc ( 60 % + 11.2px ) " but got "200 % "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [16px] to [200%] at (0.6) should be [calc(120% + 6.4px)]
  assert_equals: expected "calc ( 120 % + 6.4px ) " but got "200 % "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [1em] to [0px] at (0) should be [16px]
  assert_equals: expected "16px " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [1em] to [0px] at (0.3) should be [11.2px]
  assert_equals: expected "11.2px " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [1em] to [0px] at (0.6) should be [6.4px]
  assert_equals: expected "6.4px " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [1em] to [0px] at (0) should be [16px]
  assert_equals: expected "16px " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [1em] to [0px] at (0.3) should be [11.2px]
  assert_equals: expected "11.2px " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [1em] to [0px] at (0.6) should be [6.4px]
  assert_equals: expected "6.4px " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [1em] to [32px] at (0) should be [16px]
  assert_equals: expected "16px " but got "32px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [1em] to [32px] at (0.3) should be [20.8px]
  assert_equals: expected "20.8px " but got "32px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [1em] to [32px] at (0.6) should be [25.6px]
  assert_equals: expected "25.6px " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [1em] to [32px] at (0) should be [16px]
  assert_equals: expected "16px " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [1em] to [32px] at (0.3) should be [20.8px]
  assert_equals: expected "20.8px " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [1em] to [32px] at (0.6) should be [25.6px]
  assert_equals: expected "25.6px " but got "32px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [1em] to [0%] at (0) should be [calc(0% + 16px)]
  assert_equals: expected "calc ( 0 % + 16px ) " but got "0 % "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [1em] to [0%] at (0.3) should be [calc(0% + 11.2px)]
  assert_equals: expected "calc ( 0 % + 11.2px ) " but got "0 % "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [1em] to [0%] at (0.6) should be [calc(0% + 6.4px)]
  assert_equals: expected "calc ( 0 % + 6.4px ) " but got "0 % "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [1em] to [0%] at (0) should be [calc(0% + 16px)]
  assert_equals: expected "calc ( 0 % + 16px ) " but got "0 % "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [1em] to [0%] at (0.3) should be [calc(0% + 11.2px)]
  assert_equals: expected "calc ( 0 % + 11.2px ) " but got "0 % "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [1em] to [0%] at (0.6) should be [calc(0% + 6.4px)]
  assert_equals: expected "calc ( 0 % + 6.4px ) " but got "0 % "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [1em] to [200%] at (0) should be [calc(0% + 16px)]
  assert_equals: expected "calc ( 0 % + 16px ) " but got "200 % "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [1em] to [200%] at (0.3) should be [calc(60% + 11.2px)]
  assert_equals: expected "calc ( 60 % + 11.2px ) " but got "200 % "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [1em] to [200%] at (0.6) should be [calc(120% + 6.4px)]
  assert_equals: expected "calc ( 120 % + 6.4px ) " but got "200 % "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [1em] to [200%] at (0) should be [calc(0% + 16px)]
  assert_equals: expected "calc ( 0 % + 16px ) " but got "200 % "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [1em] to [200%] at (0.3) should be [calc(60% + 11.2px)]
  assert_equals: expected "calc ( 60 % + 11.2px ) " but got "200 % "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [1em] to [200%] at (0.6) should be [calc(120% + 6.4px)]
  assert_equals: expected "calc ( 120 % + 6.4px ) " but got "200 % "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [0px] at (0) should be [100%]
  assert_equals: expected "100 % " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [0px] at (0.3) should be [70%]
  assert_equals: expected "70 % " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [0px] at (0.6) should be [40%]
  assert_equals: expected "40 % " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [0px] at (1) should be [0%]
  assert_equals: expected "0 % " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [0px] at (0) should be [100%]
  assert_equals: expected "100 % " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [0px] at (0.3) should be [70%]
  assert_equals: expected "70 % " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [0px] at (0.6) should be [40%]
  assert_equals: expected "40 % " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [0px] at (1) should be [0%]
  assert_equals: expected "0 % " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [32px] at (0) should be [100%]
  assert_equals: expected "100 % " but got "32px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [32px] at (0.3) should be [calc(70% + 9.6px)]
  assert_equals: expected "calc ( 70 % + 9.6px ) " but got "32px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [32px] at (0.6) should be [calc(40% + 19.2px)]
  assert_equals: expected "calc ( 40 % + 19.2px ) " but got "32px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [32px] at (1) should be [calc(0% + 32px)]
  assert_equals: expected "calc ( 0 % + 32px ) " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [32px] at (0) should be [100%]
  assert_equals: expected "100 % " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [32px] at (0.3) should be [calc(70% + 9.6px)]
  assert_equals: expected "calc ( 70 % + 9.6px ) " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [32px] at (0.6) should be [calc(40% + 19.2px)]
  assert_equals: expected "calc ( 40 % + 19.2px ) " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [32px] at (1) should be [calc(0% + 32px)]
  assert_equals: expected "calc ( 0 % + 32px ) " but got "32px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [0em] at (0) should be [100%]
  assert_equals: expected "100 % " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [0em] at (0.3) should be [70%]
  assert_equals: expected "70 % " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [0em] at (0.6) should be [40%]
  assert_equals: expected "40 % " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [0em] at (1) should be [0%]
  assert_equals: expected "0 % " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [0em] at (0) should be [100%]
  assert_equals: expected "100 % " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [0em] at (0.3) should be [70%]
  assert_equals: expected "70 % " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [0em] at (0.6) should be [40%]
  assert_equals: expected "40 % " but got "0px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [0em] at (1) should be [0%]
  assert_equals: expected "0 % " but got "0px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [2em] at (0) should be [100%]
  assert_equals: expected "100 % " but got "32px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [2em] at (0.3) should be [calc(70% + 9.6px)]
  assert_equals: expected "calc ( 70 % + 9.6px ) " but got "32px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [2em] at (0.6) should be [calc(40% + 19.2px)]
  assert_equals: expected "calc ( 40 % + 19.2px ) " but got "32px "
[FAIL] CSS Transitions: property <text-decoration-thickness> from [100%] to [2em] at (1) should be [calc(0% + 32px)]
  assert_equals: expected "calc ( 0 % + 32px ) " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [2em] at (0) should be [100%]
  assert_equals: expected "100 % " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [2em] at (0.3) should be [calc(70% + 9.6px)]
  assert_equals: expected "calc ( 70 % + 9.6px ) " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [2em] at (0.6) should be [calc(40% + 19.2px)]
  assert_equals: expected "calc ( 40 % + 19.2px ) " but got "32px "
[FAIL] CSS Transitions with transition: all: property <text-decoration-thickness> from [100%] to [2em] at (1) should be [calc(0% + 32px)]
  assert_equals: expected "calc ( 0 % + 32px ) " but got "32px "
Harness: the test ran to completion.
